+++ include/Aulib/AudioDecoderWildmidi.h
@@ -60,7 +60,7 @@ public:
      *  \retval true WildMIDI was initialized sucessfully.
      *  \retval false WildMIDI could not be initialized.
      */
-    static bool init(const std::string configFile, unsigned short rate, bool hqResampling,
+    static bool init(const std::string configFile, int rate, bool hqResampling,
                      bool reverb);
 
     /*!
@@ -74,9 +74,9 @@ public:
     static void quit();
 
     bool open(SDL_RWops* rwops) override;
-    unsigned getChannels() const override;
-    unsigned getRate() const override;
-    size_t doDecoding(float buf[], size_t len, bool& callAgain) override;
+    int getChannels() const override;
+    int getRate() const override;
+    int doDecoding(float buf[], int len, bool& callAgain) override;
     bool rewind() override;
     float duration() const override;
     bool seekToTime(float seconds) override;
